+++ python/cudf/dataframe/column.py
@@ -9,6 +9,7 @@ from numbers import Number
 import numpy as np
 import pandas as pd
 from numba import cuda
+from numba.cuda.cudadrv.devicearray import DeviceNDArray
 
 from librmm_cffi import librmm as rmm
 
@@ -403,6 +404,9 @@ class Column(object):
                 return self.replace(data=newbuffer)
         elif isinstance(arg, (list, np.ndarray)):
             arg = np.array(arg)
+            arg = rmm.to_device(arg)
+
+        if isinstance(arg, DeviceNDArray):
             return self.take(arg)
         else:
             raise NotImplementedError(type(arg))